+++ coordinator/app/build.gradle
@@ -49,7 +49,6 @@ dependencies {
   implementation project(':coordinator:clients:smart-contract-client')
   implementation project(':coordinator:clients:forced-transactions-client')
   implementation project(':coordinator:ethereum:finalization-monitor')
-  implementation project(':coordinator:ethereum:common')
   implementation project(':coordinator:ethereum:models-helper')
   implementation project(':coordinator:ethereum:blob-submitter')
   implementation project(':coordinator:ethereum:message-anchoring')
