--- coordinator/ethereum/common/build.gradle
@@ -1,11 +1,0 @@
-plugins {
-  id 'net.consensys.zkevm.kotlin-library-conventions'
-}
-
-dependencies {
-  implementation(project(':coordinator:core-interfaces'))
-  implementation project(':jvm-libs:linea:web3j-extensions')
-  implementation("org.web3j:core:${libs.versions.web3j.get()}") {
-    exclude group: "org.slf4j", module: "slf4j-nop"
-  }
-}
