+++ coordinator/ethereum/finalization-monitor/build.gradle
@@ -5,7 +5,6 @@ plugins {
 
 dependencies {
   api(project(":coordinator:core"))
-  implementation(project(":coordinator:ethereum:common"))
   implementation("io.vertx:vertx-core")
   implementation("org.web3j:crypto:${libs.versions.web3j.get()}") {
     exclude group: "org.slf4j", module: "slf4j-nop"
