+++ coordinator/ethereum/test-utils/build.gradle
@@ -10,7 +10,6 @@ dependencies {
   implementation(project(":coordinator:clients:smart-contract-client"))
   implementation(project(":jvm-libs:linea:web3j-extensions"))
   implementation(testFixtures(project(":jvm-libs:linea:web3j-extensions")))
-  implementation(project(":coordinator:ethereum:common"))
   implementation(project(":jvm-libs:linea:testing:file-system"))
   implementation("org.web3j:core:${libs.versions.web3j.get()}") {
     exclude group: 'org.slf4j', module: 'slf4j-nop'
