+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/ECKeypairSignerAdapter.kt
@@ -1,11 +1,13 @@
-package net.consensys.zkevm.ethereum.signing
+package linea.web3j
 
-import linea.crypto.Signer
 import org.web3j.crypto.ECDSASignature
 import org.web3j.crypto.ECKeyPair
 import java.math.BigInteger
 
-class ECKeypairSignerAdapter(private val adaptee: Signer, publicKey: BigInteger) :
+class ECKeypairSignerAdapter(
+  private val signerDelegate: (ByteArray) -> Pair<BigInteger, BigInteger>,
+  publicKey: BigInteger,
+) :
   ECKeyPair(BigInteger.ZERO, publicKey) {
   override fun equals(other: Any?): Boolean {
     if (this === other) {
@@ -17,21 +19,21 @@ class ECKeypairSignerAdapter(private val adaptee: Signer, publicKey: BigInteger)
 
     val ecKeyPair = other as ECKeypairSignerAdapter
 
-    return super.equals(other) && adaptee == ecKeyPair.adaptee
+    return super.equals(other) && signerDelegate == ecKeyPair.signerDelegate
   }
 
   override fun hashCode(): Int {
     var result = super.hashCode()
-    result = 31 * result + adaptee.hashCode()
+    result = 31 * result + signerDelegate.hashCode()
     return result
   }
 
   override fun getPrivateKey(): BigInteger {
-    throw Exception("Key is managed by Signer adaptee, $adaptee!")
+    throw RuntimeException("Key is managed by delegated Signer: $signerDelegate!")
   }
 
   override fun sign(transaction: ByteArray): ECDSASignature {
-    val (r, s) = adaptee.sign(transaction)
+    val (r, s) = signerDelegate(transaction)
     return ECDSASignature(r, s)
   }
 }
