+++ jvm-libs/linea/web3j-extensions/src/main/kotlin/linea/web3j/Web3SignerTxSignService.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.ethereum.crypto
+package linea.web3j
 
 import org.web3j.crypto.Credentials
 import org.web3j.crypto.RawTransaction
