+++ settings.gradle
@@ -49,7 +49,6 @@ include 'coordinator:clients:shomei-client'
 include 'coordinator:clients:smart-contract-client'
 include 'coordinator:clients:web3signer-client'
 include 'coordinator:clients:forced-transactions-client'
-include 'coordinator:ethereum:common'
 include 'coordinator:ethereum:models-helper'
 include 'coordinator:ethereum:finalization-monitor'
 include 'coordinator:ethereum:forced-transactions'
